pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infrastructure-monitoring-agent"
version = "1.0.0"
description = "Automated SLA monitoring system for vessel infrastructure"
authors = [
    {name = "Fleet Operations Team"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "influxdb-client>=1.38.0",
    "apscheduler>=3.10.4",
    "jira>=3.5.2",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "flask>=3.0.0",
    "slack-sdk>=3.26.1",
    "httpx>=0.25.2",
    "requests>=2.31.0",
    "cryptography>=41.0.7",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[project.scripts]
monitoring-agent = "main:cli_main"

[project.urls]
Homepage = "https://github.com/your-org/infrastructure-monitoring-agent"
Documentation = "https://github.com/your-org/infrastructure-monitoring-agent/blob/main/README.md"
Repository = "https://github.com/your-org/infrastructure-monitoring-agent"
Issues = "https://github.com/your-org/infrastructure-monitoring-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"